Seperated the document store from the main store and added lifecycle hooks for faster component loadings
