+++ frontend/src/stores/main-store.js
@@ -13,7 +13,6 @@ export const useMainStore = defineStore("main", {
     solution: "",
     models: JSON.parse(localStorage.getItem("models")) || [],
     model_name: localStorage.getItem("model_name") || "",
-    documents: [],
     show_uploader: false,
   }),
   getters: {},
@@ -134,106 +133,6 @@ export const useMainStore = defineStore("main", {
         });
 
       //
-    },
-
-    async get_documents_list() {
-      try {
-        const response = await fetch(`${baseUrl}/documents`, {
-          method: "GET",
-          headers: {
-            Accept: "application/json",
-            Authorization: `Bearer ${authStore.token}`,
-          },
-        });
-
-        if (response.status === 401) {
-          authStore.clearToken();
-          window.location.href = "/login"; // Or use Vue Router
-          return;
-        }
-
-        if (!response.ok) {
-          // Extract error message if available
-          const errorData = await response.json().catch(() => ({})); // Parsing might fail, default to empty object
-          const errorMessage =
-            errorData.message || `Error: ${response.statusText}`;
-          throw new Error(errorMessage);
-        }
-
-        const data = await response.json();
-        this.documents = data;
-
-        //
-      } catch (error) {
-        // Handle network errors and HTTP errors
-        if (error.name === "TypeError") {
-          // This typically indicates a network error
-          console.error("Network error: Could not reach the server");
-          Notify.create({
-            color: "negative",
-            position: "bottom",
-            message: error.message,
-            icon: "report_problem",
-          });
-        } else {
-          // HTTP error, or some other error
-          console.error(`API error: ${error.message}`);
-          Notify.create({
-            color: "negative",
-            position: "bottom",
-            message: error.message,
-            icon: "report_problem",
-          });
-        }
-
-        // You can rethrow the error or handle it in some way, e.g., user notification
-      }
-    },
-
-    async upload_documents() {
-      // returning a Promise
-
-      return new Promise((resolve) => {
-        // simulating a delay of 2 seconds
-        setTimeout(() => {
-          resolve({
-            url: `${baseUrl}/documents/upload`,
-            method: "POST",
-            headers: [
-              {
-                name: "Authorization",
-                value: `Bearer ${authStore.token}`,
-              },
-            ],
-          });
-        }, 2000);
-      });
-    },
-
-    async uploaded_success() {
-      Notify.create({
-        color: "positive",
-        position: "bottom",
-        message: "uploaded",
-        icon: "done",
-      });
-
-      setTimeout(() => {
-        this.show_uploader = false;
-      }, 2000);
-
-      this.get_documents_list();
-    },
-
-    async upload_failed() {
-      // FIXME: chech if token is not valid anymore
-      Notify.create({
-        color: "negative",
-        position: "bottom",
-        message: "could not be uploaded",
-        icon: "report_problem",
-      });
-      this.show_uploader = true;
     },
   },
 });
